+++ db/model/entity/wallet.go
@@ -11,11 +11,11 @@ import (
 // Wallet is the golang structure for table wallet.
 type Wallet struct {
 	Uid           int64       `json:"uid"           description:""`
-	Balance       int64       `json:"balance"       description:"余额"`
+	Balance       float64     `json:"balance"       description:"余额"`
 	TotalBet      float64     `json:"totalBet"      description:"总投注"`
 	TotalDeposit  float64     `json:"totalDeposit"  description:"总充值"`
 	TotalWithdraw float64     `json:"totalWithdraw" description:"总提现"`
-	Freeze        int64       `json:"freeze"        description:"冻结"`
+	Freeze        float64     `json:"freeze"        description:"冻结"`
 	Account       string      `json:"account"       description:"账号"`
 	ParentPath    string      `json:"parentPath"    description:"上级路经"`
 	TotalProfit   float64     `json:"totalProfit"   description:"总盈利"`
